requirements.txt
git+https://github.com/Clari-7744/jishaku@rtfd
git+https://github.com/Clari-7744/dpy-utils
git+https://github.com/iDevision/enhanced-discord.py
aiosqlite
python-dotenv
aiohttp